enhancement/uint256-to-uint (#132)

+++ contracts/ElectoralSystems/ElectoralSystemInterface.sol
@@ -5,6 +5,6 @@ import "../OrganizationInterface.sol";
 // @todo this is ugly, lets see if we can do it with libraries later
 interface ElectoralSystemInterface {
 
-    function winner(OrganizationInterface organization, uint256 id) public view returns (uint256);
+    function winner(OrganizationInterface organization, uint id) public view returns (uint);
 
 }
