+++ contracts/Managers/VotingManager.sol
@@ -6,13 +6,13 @@ import "./VotingManagerInterface.sol";
 contract VotingManager is VotingManagerInterface, Ownable {
 
     struct Vote {
-        uint256 weight;
-        uint256 choice;
+        uint weight;
+        uint choice;
     }
 
     struct VoteData {
-        uint256 quorum;
-        mapping (uint => uint256) choices; // @todo unituitive name
+        uint quorum;
+        mapping (uint => uint) choices; // @todo unituitive name
 
         // @todo lets see if we can do this nicer, but we need it so we unvote the correct amount in case value may
         // have changed.
@@ -25,7 +25,7 @@ contract VotingManager is VotingManagerInterface, Ownable {
     /// @param proposal Id of the proposal to vote on.
     /// @param voter Address of the voter.
     /// @param choice Voters selected option.
-    function vote(uint proposal, address voter, uint choice, uint256 weight) external onlyOwner {
+    function vote(uint proposal, address voter, uint choice, uint weight) external onlyOwner {
         require(!voted(proposal, voter));
         require(weight > 0);
 
@@ -55,7 +55,7 @@ contract VotingManager is VotingManagerInterface, Ownable {
     /// @param proposal Id of the proposal.
     /// @param choice Selected option.
     /// @return count of votes for option.
-    function votes(uint proposal, uint choice) external view returns (uint256) {
+    function votes(uint proposal, uint choice) external view returns (uint) {
         return proposalVotes[proposal].choices[choice];
     }
 
@@ -69,7 +69,7 @@ contract VotingManager is VotingManagerInterface, Ownable {
     /// @dev Returns the achieved quorum for Proposal.
     /// @param proposal Id of the proposal.
     /// @return quorum of votes
-    function quorum(uint proposal) public view returns (uint256) {
+    function quorum(uint proposal) public view returns (uint) {
         return proposalVotes[proposal].quorum;
     }
 }
