+++ test/mocks/VotingPowerMock.sol
@@ -4,11 +4,11 @@ import "../../contracts/Voting/VotingPowerInterface.sol";
 
 contract VotingPowerMock is VotingPowerInterface {
 
-    function maximumQuorum() public view returns (uint256) {
+    function maximumQuorum() public view returns (uint) {
         return 3;
     }
 
-    function votingWeightOf(address, ProposalInterface) public view returns (uint256) {
+    function votingWeightOf(address, ProposalInterface) public view returns (uint) {
         return 1;
     }
 }
